+++ source/geh_common/release-notes.md
@@ -1,5 +1,37 @@
 # GEH Common Release Notes
 
+## Version 5.4.0
+
+**Subpackage**: `geh_common.testing`
+
+Added decorator `@testing` to log content of data frames returned from function invocations.
+
+Example usage:
+```python
+from geh_common.testing.dataframes import configure_testing, testing
+
+
+@testing()
+def my_function(spark: SparkSession) -> DataFrame:
+    return spark.read.parquet("data.parquet")
+
+
+if __name__ == "__main__":
+    configure_testing(True)
+    my_function()
+```
+
+Example output:
+```
+>>>In some_module.py:91: my_function(...) returned:                         
++---+----+
+|id |name|
++---+----+
+|1  |a   |
+|2  |b   |
++---+----+
+```
+
 ## Version 5.3.4
 
 - Added more valid types to `source/geh_common/src/geh_common/domain/types/`